trivial: Use the correct integer type when getting the method caller

+++ src/fu-daemon.c
@@ -234,7 +234,7 @@ fu_daemon_create_request(FuDaemon *self, const gchar *sender, GError **error)
 {
 	FuDaemonSenderItem *sender_item;
 	FwupdDeviceFlags device_flags = FWUPD_DEVICE_FLAG_NONE;
-	guint64 calling_uid = 0;
+	guint calling_uid = 0;
 	g_autoptr(FuEngineRequest) request = fu_engine_request_new(FU_ENGINE_REQUEST_KIND_ACTIVE);
 	g_autoptr(GVariant) value = NULL;
